+++ inst/doc/regression.R
@@ -1,0 +1,10 @@
+## ---- include = FALSE----------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+## ---- echo = FALSE-------------------------------------------------------
+spec_frame <- read.csv("regression.csv")
+knitr::kable(spec_frame, row.names = FALSE)
+