+++ tests/testthat.R
@@ -1,4 +1,12 @@
 library(testthat)
 library(tidypredict)
+library(xgboost)
+library(purrr)
+library(dplyr)
+library(Cubist)
+library(mlbench)
+library(rlang)
+library(partykit)
+library(yaml)
 
 test_check("tidypredict")